{{cookiecutter.project_name}}/requirements/base.txt
fastapi==0.103.2
pydantic-settings==2.0.3

hypercorn==0.14.4
python-decouple==3.8
asyncclick==8.1.3.4
email-validator==2.0.0.post2
aioredis==2.0.1

SQLAlchemy==2.0.22
asyncpg==0.28.0
aiosqlite==0.19.0
alembic==1.12.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4